requirements-dev.txt
pytest>=8.2
pytest-cov>=5.0
coverage>=7.0
duckdb
pyyaml
tqdm
colorama
requests
dash
plotly